# Grid view optimization