--- a/airflow-core/src/airflow/api_fastapi/core_api/datamodels/ui/grid.py
+++ b/airflow-core/src/airflow/api_fastapi/core_api/datamodels/ui/grid.py
@@ -21,7 +21,6 @@ from datetime import datetime
 
 from pydantic import BaseModel, Field
 
-from airflow.api_fastapi.core_api.datamodels.ui.structure import StructureDataResponse
 from airflow.utils.state import DagRunState, TaskInstanceState
 from airflow.utils.types import DagRunType
 
@@ -38,6 +37,13 @@ class GridTaskInstanceSummary(BaseModel):
     task_count: int
     state: TaskInstanceState | None
     note: str | None
+
+
+class LightGridTaskInstanceSummary(BaseModel):
+    """Task Instance Summary model for the Grid UI."""
+
+    task_id: str
+    state: TaskInstanceState | None
 
 
 class GridDAGRunwithTIs(BaseModel):
@@ -57,8 +63,15 @@ class GridDAGRunwithTIs(BaseModel):
     task_instances: list[GridTaskInstanceSummary]
 
 
+class GridTISummaries(BaseModel):
+    """DAG Run model for the Grid UI."""
+
+    run_id: str
+    dag_id: str
+    task_instances: list[LightGridTaskInstanceSummary]
+
+
 class GridResponse(BaseModel):
     """Response model for the Grid UI."""
 
     dag_runs: list[GridDAGRunwithTIs]
-    structure: StructureDataResponse