--- a/airflow-core/src/airflow/api_fastapi/core_api/datamodels/ui/grid.py
+++ b/airflow-core/src/airflow/api_fastapi/core_api/datamodels/ui/grid.py
@@ -44,9 +44,6 @@ class LightGridTaskInstanceSummary(BaseModel):
 
     task_id: str
     state: TaskInstanceState | None
-    child_states: dict[TaskInstanceState, int] | None
-    min_start_date: datetime | None
-    max_end_date: datetime | None
 
 
 class GridDAGRunwithTIs(BaseModel):