--- a/airflow-core/src/airflow/api_fastapi/core_api/services/ui/grid.py
+++ b/airflow-core/src/airflow/api_fastapi/core_api/services/ui/grid.py
@@ -18,6 +18,7 @@
 from __future__ import annotations
 
 import contextlib
+from collections import Counter
 from uuid import UUID
 
 import structlog
@@ -309,3 +310,61 @@ def _get_node_by_id(nodes, node_id):
         if node["id"] == node_id:
             return node
     return {}
+
+
+def _is_task_node_mapped_task_group(task_node: BaseOperator | MappedTaskGroup | TaskMap | None) -> bool:
+    """Check if the Task Node is a Mapped Task Group."""
+    return type(task_node) is MappedTaskGroup
+
+
+def agg_state(states):
+    states = Counter(states)
+    for state in state_priority:
+        if state in states:
+            return state
+    return "no_status"
+
+
+def _find_aggregates(
+    node: TaskGroup | BaseOperator | MappedTaskGroup | TaskMap | None,
+    parent_node: TaskGroup | BaseOperator | MappedTaskGroup | TaskMap | None,
+    ti_states: dict[str, list[str]],
+) -> dict | None:
+    """Recursively fill the Task Group Map."""
+    node_id = node.node_id
+    parent_id = parent_node.node_id if parent_node else None
+
+    if node is None:
+        return
+
+    if isinstance(node, MappedOperator):
+        yield {
+            "task_id": node_id,
+            "type": "mapped_task",
+            "parent_id": parent_id,
+            "state": agg_state(ti_states[node_id]),
+        }
+
+        return
+    if isinstance(node, TaskGroup):
+        states = []
+        for child in get_task_group_children_getter()(node):
+            for child_node in _find_aggregates(node=child, parent_node=node, ti_states=ti_states):
+                states.append(child_node["state"])
+                yield child_node
+        if node_id:
+            yield {
+                "task_id": node_id,
+                "type": "group",
+                "parent_id": parent_id,
+                "state": agg_state(states),
+            }
+        return
+    if isinstance(node, BaseOperator):
+        yield {
+            "task_id": node_id,
+            "type": "task",
+            "parent_id": parent_id,
+            "state": agg_state(ti_states[node_id]),
+        }
+        return