… airflow-core/src/airflow/api_fastapi/core_api/openapi/_private_ui.yaml
… airflow-core/src/airflow/ui/openapi-gen/queries/common.ts
… airflow-core/src/airflow/ui/openapi-gen/queries/ensureQueryData.ts
… airflow-core/src/airflow/ui/openapi-gen/queries/prefetch.ts
… airflow-core/src/airflow/ui/openapi-gen/queries/queries.ts
… airflow-core/src/airflow/ui/openapi-gen/queries/suspense.ts
… airflow-core/src/airflow/ui/openapi-gen/requests/schemas.gen.ts
… airflow-core/src/airflow/ui/openapi-gen/requests/services.gen.ts
… airflow-core/src/airflow/ui/openapi-gen/requests/types.gen.ts
… airflow-core/src/airflow/ui/src/components/DurationChart.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/DagRunSelect.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/Grid/Bar.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/Grid/Grid.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/Grid/TaskInstancesColumn.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/Grid/utils.ts
… airflow-core/src/airflow/ui/src/layouts/Details/ToggleGroups.tsx
… airflow-core/src/airflow/ui/src/layouts/Details/context.ts
… airflow-core/src/airflow/ui/src/pages/Dag/Overview/Overview.tsx
… airflow-core/src/airflow/ui/src/pages/DagRuns.tsx
… airflow-core/src/airflow/ui/src/pages/GroupTaskInstance/GroupTaskInstance.tsx
… airflow-core/src/airflow/ui/src/pages/GroupTaskInstance/Header.tsx
… airflow-core/src/airflow/ui/src/pages/Task/Task.tsx
… airflow-core/src/airflow/ui/src/queries/useGrid.ts
… airflow-core/src/airflow/ui/src/queries/useGridRuns.ts
… airflow-core/src/airflow/ui/src/queries/useGridStructure.ts
… airflow-core/src/airflow/ui/src/queries/useGridTISummaries.ts
… airflow-core/src/airflow/ui/src/queries/useRefreshOnNewDagRuns.ts
… airflow-core/src/airflow/ui/src/utils/groupTask.ts
… airflow-core/src/airflow/utils/task_group.py
… airflow-core/tests/unit/api_fastapi/core_api/routes/ui/test_grid.py